+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v5.19.0
+
 ### Added
 
 - Add `query` and `scopes` arguments to the `@can` directive https://github.com/nuwave/lighthouse/issues/1916